+++ lib/phoenix_live_view/utils.ex
@@ -63,7 +63,7 @@ defmodule Phoenix.LiveView.Utils do
   end
 
   defp configure_assigns(assigns, view, action, flash) do
-    Map.merge(assigns, %{live_view_module: view, live_view_action: action, flash: flash})
+    Map.merge(assigns, %{live_module: view, live_action: action, flash: flash})
   end
 
   @doc """